feat: add comprehensive integration tests for data transformation pipeline
