+++ backend/tests/integration/data-transformation-pipeline.test.ts
@@ -1,0 +1,33 @@
+import { describe } from '@jest/globals';
+
+/**
+ * Data Transformation Pipeline Integration Tests
+ * 
+ * This test suite covers the complete data transformation pipeline:
+ * DOM extraction → Result transformation → Final API response
+ * 
+ * The pipeline ensures that:
+ * 1. DOM extractors include 'url' field for backend validation
+ * 2. Result transformers properly strip 'url' field for frontend API
+ * 3. The unified Song interface is maintained throughout the pipeline
+ * 
+ * This comprehensive test suite would have caught the song-only search 
+ * failure during the frontend-backend unification.
+ * 
+ * Test files are organized by functionality:
+ * - search-results-pipeline.test.ts: Search result processing
+ * - artist-songs-pipeline.test.ts: Artist-specific song processing  
+ * - unified-interface-validation.test.ts: Cross-source consistency
+ * - edge-cases.test.ts: Edge cases and error scenarios
+ */
+
+describe('Data Transformation Pipeline Integration', () => {
+  // Individual test files are automatically discovered by Jest
+  // This file serves as documentation and organization
+});
+
+// Import all test modules to ensure they run as part of this suite
+import './data-transformation-pipeline/search-results-pipeline.test.js';
+import './data-transformation-pipeline/artist-songs-pipeline.test.js';
+import './data-transformation-pipeline/unified-interface-validation.test.js';
+import './data-transformation-pipeline/edge-cases.test.js';
